Adds support for event field
Couple of folks gave me feedback that they set an event field in
Riemann when sending to Alerta and other destinations. The current
webhook discards this potential field and uses service instead. This
changes the behavior so that if the event field exists it is used
otherwise it defaults to service.