Fix Stackdriver webhook
This commit fixes two minor but breaking issues with the stackdriver webhook view:
1. The parse_stackdriver method was returning a tuple of a state (string) and an Alert,
where calling code was expecting it to return only an Alert.
2. The old code assumed that Stackdriver was setting the header
"Content-Type: application/json". Stackdriver should indeed set this, but
in practice it doesn't, so using request.get_json(force=True)
instead of request.json is an easy workaround.